+++ R/phase2.r
@@ -106,6 +106,7 @@ proc2subphase <- function(z, x, subphase, ...)
 		## cat(z$thav, z$theta, '\n')
 		z$prod0 <- rep(0, z$pp)
 		z$prod1 <- rep(0, z$pp)
+		functionAv <- 0
 		## ###############################################
 		## do the iterations for this repeat of this subphase
 		## ##############################################
@@ -227,6 +228,7 @@ doIterations<- function(z, x, subphase,...)
 	if (z$returnThetas)
 	{
 		thetas <- NULL
+		sfs <- NULL
 	}
 	repeat
 	{
@@ -258,14 +260,31 @@ doIterations<- function(z, x, subphase,...)
 		}
 		zsmall$nit <- z$nit
 		if (x$dolby) {zsmall$Deriv <- TRUE} ## include scores in FRAN
+		
+
+#Report(paste("z$theta: ", "\n"), cf)
+#PrtOutMat(as.matrix(z$targets), cf)
+#PrtOutMat(as.matrix(z$theta), cf)
+#PrtOutMat(as.matrix(zsmall$theta), cf)
+
+
 		if (z$int == 1) ## then no parallel runs at this level
 		{
 			zz <- x$FRAN(zsmall, xsmall)
 			fra <- colSums(zz$fra) - z$targets
+			
+			
+#Report(paste("fra (1): ", "\n"), cf)
+#PrtOutMat(as.matrix(fra), cf)
+
 			if (!zz$OK)
 			{
 				z$OK <- zz$OK
 				break
+			}
+			if (z$returnThetas)
+			{
+				sfs <- rbind(sfs, c(subphase, fra))
 			}
 			if (x$dolby)
 				## subtract regression on scores;
@@ -398,6 +417,10 @@ doIterations<- function(z, x, subphase,...)
 			sumfra <- sumfra + fra
 			fra <- sumfra
 		}
+
+#Report(paste("fra: ", "\n"), cf)
+#PrtOutMat(as.matrix(fra), cf)
+
 		if (x$standardizeVar)
 		{
 			if (x$diagg)
@@ -460,30 +483,23 @@ doIterations<- function(z, x, subphase,...)
 		z$positivized[fchange > z$theta] <- z$positivized[fchange > z$theta] +1
 		z$positivized[!z$posj] <- 0
 		fchange <- ifelse(z$posj & (fchange > z$theta), z$theta * 0.5, fchange)
-		# make step
+		# make update step
 		if (subphase > x$doubleAveraging)
 		{
 			zsmall$theta <- (z$thav/z$thavn) - fchange
 		}
 		else
 		{
-			zsmall$theta <- zsmall$theta - fchange
+				zsmall$theta <- zsmall$theta - fchange
 		}
 		z$theta <- zsmall$theta
-		if (!is.null(z$thetaStore))
-		{
-			if (z$nit <= dim(z$thetaStore)[1])
-			{
-				z$thetaStore[z$nit,] <- z$theta
-			}
-			else
-			{
-				message('thetaStore?')
-				browser()
-			}
-		}
 		z$thav <- z$thav + zsmall$theta
 		z$thavn <- z$thavn + 1
+ 
+#Report(paste("thavs: ", round(z$thavn), "\n"), cf)
+#PrtOutMat(as.matrix(z$thav), cf)
+
+
 		if (x$maxlike && !is.null(x$moreUpdates) && x$moreUpdates > 0)
 		{
 			z <- doMoreUpdates(z, x, x$moreUpdates * subphase)
@@ -510,6 +526,7 @@ doIterations<- function(z, x, subphase,...)
 	if (z$returnThetas)
 	{
 		z$thetas <- rbind(z$thetas, thetas)
+		z$sfs <- rbind(z$sfs, sfs)
 	}
 	z
 }