+++ R/sienaGOF.r
@@ -983,18 +983,18 @@ sparseMatrixExtraction <-
 	dimsOfDepVar<- attr(obsData[[groupName]]$depvars[[varName]], "netdims")
 	if (attr(obsData[[groupName]]$depvars[[varName]], "missing"))
 	{
-	if (attr(obsData[[groupName]]$depvars[[varName]], "sparse"))
-	{
-		missings <-
-			(is.na(obsData[[groupName]]$depvars[[varName]][[period]]) |
-			is.na(obsData[[groupName]]$depvars[[varName]][[period+1]]))*1
-	}
-	else
-	{
-		missings <- Matrix(
-			(is.na(obsData[[groupName]]$depvars[[varName]][,,period]) |
-			is.na(obsData[[groupName]]$depvars[[varName]][,,period+1]))*1)
-	}
+		if (attr(obsData[[groupName]]$depvars[[varName]], "sparse"))
+		{
+			missings <-
+				(is.na(obsData[[groupName]]$depvars[[varName]][[period]]) |
+				is.na(obsData[[groupName]]$depvars[[varName]][[period+1]]))*1
+		}
+		else
+		{
+			missings <- Matrix(
+				(is.na(obsData[[groupName]]$depvars[[varName]][,,period]) |
+				is.na(obsData[[groupName]]$depvars[[varName]][,,period+1]))*1)
+		}
 	}
 	if (is.null(i))
 	{
@@ -1011,8 +1011,8 @@ sparseMatrixExtraction <-
 			if (attr(obsData[[groupName]]$depvars[[varName]], "structural"))
 			{
 				extractedMatrix <- changeToStructural(extractedMatrix,
-				Matrix(obsData[[groupName]]$depvars[[varName]][[period]]))
-		}
+					Matrix(obsData[[groupName]]$depvars[[varName]][[period]]))
+			}
 		}
 		else # not sparse
 		{
@@ -1023,7 +1023,7 @@ sparseMatrixExtraction <-
 			{
 				extractedMatrix <- changeToStructural(extractedMatrix,
 				Matrix(obsData[[groupName]]$depvars[[varName]][,,period]))
-		}
+			}
 		}
 		if(!isBipartite){ diag(extractedMatrix) <- 0} # not guaranteed by data input
 	}
@@ -1034,24 +1034,24 @@ sparseMatrixExtraction <-
 				sims[[i]][[groupName]][[varName]][[period]][,1],
 				sims[[i]][[groupName]][[varName]][[period]][,2],
 				x=sims[[i]][[groupName]][[varName]][[period]][,3],
-				dims=dimsOfDepVar[1:2] )
+				dims=dimsOfDepVar[1:2], repr="T")
 		if (attr(obsData[[groupName]]$depvars[[varName]], "structural"))
 		{
 		# If observation at end of period contains structural values
 		# use these to replace the simulations.
-		if (attr(obsData[[groupName]]$depvars[[varName]], "sparse"))
-		{
+			if (attr(obsData[[groupName]]$depvars[[varName]], "sparse"))
+			{
 				extractedMatrix <- changeToNewStructural(extractedMatrix,
 				Matrix(obsData[[groupName]]$depvars[[varName]][[period]]),
 				Matrix(obsData[[groupName]]$depvars[[varName]][[period+1]]))
-		}
-		else # not sparse
-		{
+			}
+			else # not sparse
+			{
 				extractedMatrix <- changeToNewStructural(extractedMatrix,
 				Matrix(obsData[[groupName]]$depvars[[varName]][,,period]),
 				Matrix(obsData[[groupName]]$depvars[[varName]][,,period+1]))
+			}
 		}
-	}
 	}
 	## Zero missings (the 1* turns the logical into numeric):
 	if (attr(obsData[[groupName]]$depvars[[varName]], "missing"))
@@ -1084,6 +1084,7 @@ sparseMatrixExtraction0 <-
 			extractedValue[is.na(extractedValue)] <- 0
 		}
 		diag(extractedValue) <- 0 # not guaranteed by data input
+		extractedValue <- as(drop0(extractedValue), "sparseMatrix")
 	}
 	else
 	{
@@ -1092,7 +1093,7 @@ sparseMatrixExtraction0 <-
 				sims[[i]][[groupName]][[varName]][[period]][,1],
 				sims[[i]][[groupName]][[varName]][[period]][,2],
 				x=1,
-				dims=dimsOfDepVar[1:2] )
+				dims=dimsOfDepVar[1:2], repr="T")
 	}
 	extractedValue
 }
@@ -1116,7 +1117,7 @@ networkExtraction <- function (i, obsData, sims, period, groupName, varName){
 	# the number of actors (rows) plus the number of events (columns)
 	# with all actors preceding all events.
 	# Therefore the bipartiteOffset will come in handy:
-	bipartiteOffset <- ifelse (isbipartite, 1 + dimsOfDepVar[1], 1)
+	bipartiteOffset <- ifelse(isbipartite, dimsOfDepVar[1], 0)
 
 	# Initialize empty networks:
 	if (isbipartite)
@@ -1131,24 +1132,15 @@ networkExtraction <- function (i, obsData, sims, period, groupName, varName){
 	# Use what was defined in the function above:
 	matrixNetwork <- sparseMatrixExtraction(i, obsData, sims,
 						period, groupName, varName)
-	sparseMatrixNetwork <- as(matrixNetwork, "dgTMatrix")
-# For dgTMatrix, slots i and j are the rows and columns,
-# numbered from 0 to dimension - 1. Slot x are the values.
-# Actors in class network are numbered starting from 1.
-# Hence 1 must be added to missings@i and missings@j.
-# sparseMatrixNetwork@x is a column of ones;
-# the 1 in the 3d column of cbind below is redundant
-# because of the default ignore.eval=TRUE in network.edgelist.
-# But it is good to be explicit.
 	if (sum(matrixNetwork) <= 0) # else network.edgelist() below will not work
 	{
 		extractedValue <- emptyNetwork
 	}
 	else
 	{
+		tripEV <- mat2triplet(matrixNetwork)
 		extractedValue <- network::network.edgelist(
-					cbind(sparseMatrixNetwork@i + 1,
-					sparseMatrixNetwork@j + bipartiteOffset, 1),
+					cbind(tripEV$i, tripEV$j + bipartiteOffset, 1),
 					emptyNetwork)
 	}
 	extractedValue